Modularize methods that output doc pages and examples
# Conflicts:
#	tools/gulp/docs/generatePage.js
